pyproject.toml
[project]
name = "iri-api-python"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "humps>=0.2.2",
    "gql[httpx]>=3.5.0",
    "pyjwt[crypto]>=2.8.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

